% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_bounds.R
\name{attrition_bound_ate}
\alias{attrition_bound_ate}
\title{Compute bounds on ATE}
\usage{
attrition_bound_ate(Yobs, Dtr, R, ascore, zeta, is_discrete)
}
\arguments{
\item{is_discrete}{A boolean argument if the outcome is a discrete variable or not.}
}
\description{
Compute bounds on ATE
}
\keyword{internal}
