% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_bounds.R
\name{attrition_bound_ate_boot}
\alias{attrition_bound_ate_boot}
\title{Compute Variance of ATE Bound via Bootstrap}
\usage{
attrition_bound_ate_boot(Yobs, Dtr, R, ascore, zeta, n_boot = 500)
}
\arguments{
\item{Yobs}{A vector of outcomes. It should contain at least two `NA` values for each condition.}

\item{Dtr}{A vector of binary treatment indicator.}

\item{R}{A vector of missing value indicator.}

\item{ascore}{A vector of estimated attrition score.}

\item{zeta}{A vector of the sensitivity parameter. The value should be larger than 1.}
}
\description{
Compute Variance of ATE Bound via Bootstrap
}
\keyword{internal}
