% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_bounds.R
\name{attrition_bound_ate_boot_full}
\alias{attrition_bound_ate_boot_full}
\title{Compute Varinace of ATE Bounds via Full Bootstrap}
\usage{
attrition_bound_ate_boot_full(
  formula,
  varname_treat,
  Y,
  R,
  data,
  cbps,
  zeta,
  n_boot,
  is_discrete
)
}
\description{
Compute Varinace of ATE Bounds via Full Bootstrap
}
