% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_bounds.R
\name{attrition_bound_qte}
\alias{attrition_bound_qte}
\title{Estimate Bounds for QTE}
\usage{
attrition_bound_qte(Yobs, Dtr, R, ascore, zeta, probs)
}
\description{
Estimate Bounds for QTE
}
\keyword{internal}
