% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_bounds.R
\name{attrition_bound_qte_boot}
\alias{attrition_bound_qte_boot}
\title{Estimate Variance of Bounds for QTE with Bootstrap}
\usage{
attrition_bound_qte_boot(Yobs, Dtr, R, ascore, zeta, probs, n_boot)
}
\description{
Estimate Variance of Bounds for QTE with Bootstrap
}
\keyword{internal}
