% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{propensity_score}
\alias{propensity_score}
\title{Estimate propensity score}
\usage{
propensity_score(formula, data, est_ps, cbps)
}
\arguments{
\item{formula}{A formula \code{treatment ~ x1 + x2}.}

\item{data}{A data frame.}

\item{est_ps}{A boolean argument if the propensity score is estimated.
If set \code{FALSE}, empirical fraction of treated units is used.}
}
\value{
A vector of propensity score with length n (the number of observations).
}
\description{
Estimate propensity score
}
\keyword{internal}
