% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{attrition}
\alias{attrition}
\title{Estimate Average Treatment Effect under Differential Attrition}
\usage{
attrition(formula, data, estimator = "dr", est_ps = FALSE)
}
\arguments{
\item{formula}{A formula of the form \code{outcome | treatmet ~  x1 + x2},
where \code{outcoem} is the outcome of interest with missing value \code{NA}.}

\item{data}{A data frame.}

\item{estimator}{An estimator used for estimating ATE. Default is \code{"dr"} (double robust).}
}
\description{
Estimate attrition scores and estimate ATE under differential attrition
}
