% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_bounds.R
\name{attrition_bound}
\alias{attrition_bound}
\title{Estimate Bounds on Causal Effect}
\usage{
attrition_bound(
  formula,
  data,
  qoi = "ate",
  cbps = TRUE,
  zeta = c(1, 1.1, 1.2),
  probs = c(0.25, 0.5, 0.75)
)
}
\arguments{
\item{formula}{A formula of the form \code{outcome | treatmet ~  x1 + x2},
where \code{outcoem} is the outcome of interest
with missing value \code{NA}.
\code{treatment} is the binary treatment indicator.
Treatment and covariates should not have missing values.}

\item{data}{A data frame.
It should contain all variables specified in the \code{formula}.}

\item{qoi}{Quantity of interest. Currently supports the average treatment efeect (`"ate"`) or
the quantile treatment effect (`"qte"`)}

\item{cbps}{A boolean argument. If \code{TRUE}, the covariate balancing propensity
score (CBPS) is used to estimate the attrition score (and the propensity
score when \code{est_ps = TRUE}). Otherwise, \code{glm} is used.}

\item{zeta}{A vector of sensitivity parameters. The values should be greater than or equal to 1.}

\item{probs}{A vector of quantiles. The value should be between 0 and 1.
This argument is only meaninful when the quantile treatment effeect (`qoi = "qte"`)
is computed.}
}
\description{
Compute the bounds on casual effects with sensitivity parameter zeta.
}
