% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{attrition_score}
\alias{attrition_score}
\title{Estimate Attrition Score}
\usage{
attrition_score(formula, varname_treat, data)
}
\arguments{
\item{formula}{A formula of the form \code{response ~ x1 + x2}.}

\item{varname_treat}{A character string of treatment variable name in \code{data}}

\item{data}{A data frame.}
}
\description{
Estimate Attrition Score
}
\keyword{internal}
